requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
websockets==15.0.1
